+++ src/client/src/commons/form/borehole/completion/backfill.jsx
@@ -1,6 +1,5 @@
-import React from "react";
-import { getBackfills, addBackfill, updateBackfill, deleteBackfill } from "../../../../api/fetchApiV2";
-import { DataCards } from "../../../../components/dataCard/dataCards";
+import { getBackfills } from "../../../../api/fetchApiV2";
+import DataCards from "../../../../components/dataCard/dataCards";
 import BackfillInput from "./backfillInput";
 import BackfillDisplay from "./backfillDisplay";
 import { sortByDepth } from "../../../sorter.jsx";
@@ -11,9 +10,6 @@ const Backfill = ({ isEditable, completionId }) => {
       isEditable={isEditable}
       parentId={completionId}
       getData={getBackfills}
-      addData={addBackfill}
-      updateData={updateBackfill}
-      deleteData={deleteBackfill}
       cyLabel="backfill"
       addLabel="addFilling"
       emptyLabel="msgFillingEmpty"
@@ -31,5 +27,4 @@ const Backfill = ({ isEditable, completionId }) => {
     />
   );
 };
-const MemoizedBackfill = React.memo(Backfill);
-export default MemoizedBackfill;
+export default Backfill;