+++ src/client/src/commons/form/borehole/hydrogeology/fieldMeasurement.jsx
@@ -1,12 +1,7 @@
-import {
-  getFieldMeasurements,
-  addFieldMeasurement,
-  updateFieldMeasurement,
-  deleteFieldMeasurement,
-} from "../../../../api/fetchApiV2";
+import { getFieldMeasurements } from "../../../../api/fetchApiV2";
 import FieldMeasurementInput from "./fieldMeasurementInput";
 import FieldMeasurementDisplay from "./fieldMeasurementDisplay";
-import { DataCards } from "../../../../components/dataCard/dataCards";
+import DataCards from "../../../../components/dataCard/dataCards";
 import { sortByDepth } from "../../../sorter.jsx";
 
 const FieldMeasurement = ({ isEditable, boreholeId }) => {
@@ -15,9 +10,6 @@ const FieldMeasurement = ({ isEditable, boreholeId }) => {
       isEditable={isEditable}
       parentId={boreholeId}
       getData={getFieldMeasurements}
-      addData={addFieldMeasurement}
-      updateData={updateFieldMeasurement}
-      deleteData={deleteFieldMeasurement}
       cyLabel="fieldMeasurement"
       addLabel="addFieldmeasurement"
       emptyLabel="msgFieldMeasurementsEmpty"