+++ src/client/src/commons/form/borehole/hydrogeology/fieldMeasurementInput.jsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { InputAdornment, Stack } from "@mui/material";
 import { FormInput, FormSelect } from "../../../../components/form/form";
 import { DataInputCard } from "../../../../components/dataCard/dataInputCard";
-import { useDomains } from "../../../../api/fetchApiV2";
+import { addFieldMeasurement, updateFieldMeasurement, useDomains } from "../../../../api/fetchApiV2";
 import { useTranslation } from "react-i18next";
 import ObservationInput from "./observationInput";
 import { ObservationType } from "./observationType";
@@ -10,7 +10,7 @@ import { hydrogeologySchemaConstants } from "./hydrogeologySchemaConstants";
 import { FieldMeasurementParameterUnits } from "./parameterUnits";
 
 const FieldMeasurementInput = props => {
-  const { item, setSelected, parentId, addData, updateData } = props;
+  const { item, parentId } = props;
   const domains = useDomains();
   const { i18n } = useTranslation();
   const [parameterId, setParameterId] = useState(null);
@@ -34,9 +34,8 @@ const FieldMeasurementInput = props => {
   return (
     <DataInputCard
       item={item}
-      setSelected={setSelected}
-      addData={addData}
-      updateData={updateData}
+      addData={addFieldMeasurement}
+      updateData={updateFieldMeasurement}
       prepareFormDataForSubmit={prepareFormDataForSubmit}>
       <ObservationInput observation={item} boreholeId={parentId} />
       <Stack direction="row" sx={{ paddingTop: "10px" }}>