+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 - Reusable form components.
 - Quality attribute for stratigraphy.
 - Support Amazon Cognito logout flow.
-- IsOpenBorehole option for instrumentation and backfill
+- IsOpenBorehole option for instrumentation and backfill.
+- Show prompt for unsaved changes when switching data card.
 
 ### Changed